fix(ci): restore clippy baseline
