+++ src/core/providers/azure/mod.rs
@@ -370,8 +370,8 @@ mod tests {
             if read_http_headers(&mut socket).await.is_err() {
                 return;
             }
-            if socket.write_all(response.as_bytes()).await.is_err() {
-                return;
+            if let Err(err) = socket.write_all(response.as_bytes()).await {
+                eprintln!("test server failed to write response: {err}");
             }
         });
