+++ src/core/providers/sagemaker/sigv4.rs
@@ -71,7 +71,7 @@ impl SagemakerSigV4Signer {
 
         // Sort headers by key (case-insensitive)
         let mut sorted_headers: Vec<_> = canonical_headers.iter().collect();
-        sorted_headers.sort_by(|a, b| a.0.to_lowercase().cmp(&b.0.to_lowercase()));
+        sorted_headers.sort_by_key(|(key, _)| key.to_lowercase());
 
         // Build canonical headers string
         let canonical_headers_str = sorted_headers
